requirements.txt
# Core dependencies for the Windows Assistant
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-llms-groq>=0.1.0
python-dotenv>=1.0.0

# Document processing
PyMuPDF>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# App opening functionality
AppOpener>=1.7

# Desktop UI dependencies (tkinter is built into Python)
# No additional UI dependencies needed for tkinter

# Additional utilities
pathlib2>=2.3.0
